project/requirements.txt
Pillow==6.1.0
boto3==1.9.213
picamera==1.13
